Refactor CLI class
