+++ includes/CLI/Abstract_Plugin_Check_Command.php
@@ -1,0 +1,202 @@
+<?php
+/**
+ * Class WordPress\Plugin_Check\CLI\Abstract_Plugin_Check_Command
+ *
+ * @package plugin-check
+ */
+
+namespace WordPress\Plugin_Check\CLI;
+
+use WordPress\Plugin_Check\Checker\Runtime_Check;
+use WordPress\Plugin_Check\Plugin_Context;
+use WP_CLI;
+
+/**
+ * Abstract Plugin check command.
+ */
+class Abstract_Plugin_Check_Command {
+
+	/**
+	 * Plugin context.
+	 *
+	 * @since n.e.x.t
+	 * @var Plugin_Context
+	 */
+	protected $plugin_context;
+
+	/**
+	 * Output format type.
+	 *
+	 * @since n.e.x.t
+	 * @var string[]
+	 */
+	protected $output_formats = array(
+		'table',
+		'csv',
+		'json',
+	);
+
+	/**
+	 * Constructor.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param Plugin_Context $plugin_context Plugin context.
+	 */
+	public function __construct( Plugin_Context $plugin_context ) {
+		$this->plugin_context = $plugin_context;
+	}
+
+	/**
+	 * Validates the associative arguments.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param array $assoc_args List of the associative arguments.
+	 * @param array $defaults   List of the default arguments.
+	 * @return array List of the associative arguments.
+	 *
+	 * @throws WP_CLI\ExitException Show error if invalid format argument.
+	 */
+	protected function get_options( $assoc_args, $defaults ) {
+		$options = wp_parse_args( $assoc_args, $defaults );
+
+		if ( ! in_array( $options['format'], $this->output_formats, true ) ) {
+			WP_CLI::error(
+				sprintf(
+					// translators: 1. Output formats.
+					__( 'Invalid format argument, valid value will be one of [%1$s]', 'plugin-check' ),
+					implode( ', ', $this->output_formats )
+				)
+			);
+		}
+
+		return $options;
+	}
+
+	/**
+	 * Gets the formatter instance to format check results.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param array $assoc_args     Associative arguments.
+	 * @param array $default_fields Default fields.
+	 * @return WP_CLI\Formatter The formatter instance.
+	 */
+	protected function get_formatter( $assoc_args, $default_fields ) {
+		if ( isset( $assoc_args['fields'] ) ) {
+			$default_fields = wp_parse_args( $assoc_args['fields'], $default_fields );
+		}
+
+		return new WP_CLI\Formatter(
+			$assoc_args,
+			$default_fields
+		);
+	}
+
+	/**
+	 * Flattens and combines the given associative array of file errors and file warnings into a two-dimensional array.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param array $file_errors   Errors from a Check_Result, for a specific file.
+	 * @param array $file_warnings Warnings from a Check_Result, for a specific file.
+	 * @return array Combined file results.
+	 *
+	 * @SuppressWarnings(PHPMD.NPathComplexity)
+	 */
+	protected function flatten_file_results( $file_errors, $file_warnings ) {
+		$file_results = array();
+
+		foreach ( $file_errors as $line => $line_errors ) {
+			foreach ( $line_errors as $column => $column_errors ) {
+				foreach ( $column_errors as $column_error ) {
+
+					$file_results[] = array_merge(
+						$column_error,
+						array(
+							'type'   => 'ERROR',
+							'line'   => $line,
+							'column' => $column,
+						)
+					);
+				}
+			}
+		}
+
+		foreach ( $file_warnings as $line => $line_warnings ) {
+			foreach ( $line_warnings as $column => $column_warnings ) {
+				foreach ( $column_warnings as $column_warning ) {
+
+					$file_results[] = array_merge(
+						$column_warning,
+						array(
+							'type'   => 'WARNING',
+							'line'   => $line,
+							'column' => $column,
+						)
+					);
+				}
+			}
+		}
+
+		usort(
+			$file_results,
+			static function ( $a, $b ) {
+				if ( $a['line'] < $b['line'] ) {
+					return -1;
+				}
+				if ( $a['line'] > $b['line'] ) {
+					return 1;
+				}
+				if ( $a['column'] < $b['column'] ) {
+					return -1;
+				}
+				if ( $a['column'] > $b['column'] ) {
+					return 1;
+				}
+				return 0;
+			}
+		);
+
+		return $file_results;
+	}
+
+	/**
+	 * Displays the results.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param WP_CLI\Formatter $formatter Formatter class.
+	 * @param array            $results   Results.
+	 * @param string           $heading   Heading.
+	 */
+	protected function display_results( $formatter, $results, $heading = '' ) {
+		if ( ! empty( $heading ) ) {
+			WP_CLI::line( $heading );
+		}
+
+		$formatter->display_items( $results );
+
+		WP_CLI::line();
+		WP_CLI::line();
+	}
+
+	/**
+	 * Checks for a Runtime_Check in a list of checks.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @param array $checks An array of Check instances.
+	 * @return bool True if a Runtime_Check exists in the array, false if not.
+	 */
+	protected function has_runtime_check( array $checks ) {
+		foreach ( $checks as $check ) {
+			if ( $check instanceof Runtime_Check ) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+}
