+++ includes/CLI/Plugin_Check_Command.php
@@ -12,47 +12,14 @@ use WordPress\Plugin_Check\Checker\Check_Categories;
 use WordPress\Plugin_Check\Checker\Check_Repository;
 use WordPress\Plugin_Check\Checker\CLI_Runner;
 use WordPress\Plugin_Check\Checker\Default_Check_Repository;
-use WordPress\Plugin_Check\Checker\Runtime_Check;
 use WordPress\Plugin_Check\Checker\Runtime_Environment_Setup;
-use WordPress\Plugin_Check\Plugin_Context;
 use WordPress\Plugin_Check\Utilities\Plugin_Request_Utility;
 use WP_CLI;
 
 /**
  * Plugin check command.
  */
-final class Plugin_Check_Command {
-
-	/**
-	 * Plugin context.
-	 *
-	 * @since n.e.x.t
-	 * @var Plugin_Context
-	 */
-	protected $plugin_context;
-
-	/**
-	 * Output format type.
-	 *
-	 * @since n.e.x.t
-	 * @var string[]
-	 */
-	protected $output_formats = array(
-		'table',
-		'csv',
-		'json',
-	);
-
-	/**
-	 * Constructor.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param Plugin_Context $plugin_context Plugin context.
-	 */
-	public function __construct( Plugin_Context $plugin_context ) {
-		$this->plugin_context = $plugin_context;
-	}
+final class Plugin_Check_Command extends Abstract_Plugin_Check_Command {
 
 	/**
 	 * Runs plugin check.
@@ -225,13 +192,13 @@ final class Plugin_Check_Command {
 				unset( $warnings[ $file_name ] );
 			}
 			$file_results = $this->flatten_file_results( $file_errors, $file_warnings );
-			$this->display_results( $formatter, $file_name, $file_results );
+			$this->display_results( $formatter, $file_results, sprintf( 'FILE: %s', $file_name ) );
 		}
 
 		// If there are any files left with only warnings, print those next.
 		foreach ( $warnings as $file_name => $file_warnings ) {
 			$file_results = $this->flatten_file_results( array(), $file_warnings );
-			$this->display_results( $formatter, $file_name, $file_results );
+			$this->display_results( $formatter, $file_results, sprintf( 'FILE: %s', $file_name ) );
 		}
 	}
 
@@ -319,7 +286,7 @@ final class Plugin_Check_Command {
 		);
 
 		// Display results.
-		$formatter->display_items( $all_checks );
+		$this->display_results( $formatter, $all_checks );
 	}
 
 	/**
@@ -371,7 +338,7 @@ final class Plugin_Check_Command {
 		);
 
 		// Display results.
-		$formatter->display_items( $categories );
+		$this->display_results( $formatter, $categories );
 	}
 
 	/**
@@ -395,53 +362,6 @@ final class Plugin_Check_Command {
 		}
 
 		return $categories;
-	}
-
-	/**
-	 * Validates the associative arguments.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param array $assoc_args List of the associative arguments.
-	 * @param array $defaults   List of the default arguments.
-	 * @return array List of the associative arguments.
-	 *
-	 * @throws WP_CLI\ExitException Show error if invalid format argument.
-	 */
-	private function get_options( $assoc_args, $defaults ) {
-		$options = wp_parse_args( $assoc_args, $defaults );
-
-		if ( ! in_array( $options['format'], $this->output_formats, true ) ) {
-			WP_CLI::error(
-				sprintf(
-					// translators: 1. Output formats.
-					__( 'Invalid format argument, valid value will be one of [%1$s]', 'plugin-check' ),
-					implode( ', ', $this->output_formats )
-				)
-			);
-		}
-
-		return $options;
-	}
-
-	/**
-	 * Gets the formatter instance to format check results.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param array $assoc_args     Associative arguments.
-	 * @param array $default_fields Default fields.
-	 * @return WP_CLI\Formatter The formatter instance.
-	 */
-	private function get_formatter( $assoc_args, $default_fields ) {
-		if ( isset( $assoc_args['fields'] ) ) {
-			$default_fields = wp_parse_args( $assoc_args['fields'], $default_fields );
-		}
-
-		return new WP_CLI\Formatter(
-			$assoc_args,
-			$default_fields
-		);
 	}
 
 	/**
@@ -472,114 +392,5 @@ final class Plugin_Check_Command {
 		}
 
 		return $default_fields;
-	}
-
-	/**
-	 * Flattens and combines the given associative array of file errors and file warnings into a two-dimensional array.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param array $file_errors   Errors from a Check_Result, for a specific file.
-	 * @param array $file_warnings Warnings from a Check_Result, for a specific file.
-	 * @return array Combined file results.
-	 *
-	 * @SuppressWarnings(PHPMD.NPathComplexity)
-	 */
-	private function flatten_file_results( $file_errors, $file_warnings ) {
-		$file_results = array();
-
-		foreach ( $file_errors as $line => $line_errors ) {
-			foreach ( $line_errors as $column => $column_errors ) {
-				foreach ( $column_errors as $column_error ) {
-
-					$file_results[] = array_merge(
-						$column_error,
-						array(
-							'type'   => 'ERROR',
-							'line'   => $line,
-							'column' => $column,
-						)
-					);
-				}
-			}
-		}
-
-		foreach ( $file_warnings as $line => $line_warnings ) {
-			foreach ( $line_warnings as $column => $column_warnings ) {
-				foreach ( $column_warnings as $column_warning ) {
-
-					$file_results[] = array_merge(
-						$column_warning,
-						array(
-							'type'   => 'WARNING',
-							'line'   => $line,
-							'column' => $column,
-						)
-					);
-				}
-			}
-		}
-
-		usort(
-			$file_results,
-			static function ( $a, $b ) {
-				if ( $a['line'] < $b['line'] ) {
-					return -1;
-				}
-				if ( $a['line'] > $b['line'] ) {
-					return 1;
-				}
-				if ( $a['column'] < $b['column'] ) {
-					return -1;
-				}
-				if ( $a['column'] > $b['column'] ) {
-					return 1;
-				}
-				return 0;
-			}
-		);
-
-		return $file_results;
-	}
-
-	/**
-	 * Displays the results.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param WP_CLI\Formatter $formatter    Formatter class.
-	 * @param string           $file_name    File name.
-	 * @param array            $file_results Results.
-	 */
-	private function display_results( $formatter, $file_name, $file_results ) {
-		WP_CLI::line(
-			sprintf(
-				'FILE: %s',
-				$file_name
-			)
-		);
-
-		$formatter->display_items( $file_results );
-
-		WP_CLI::line();
-		WP_CLI::line();
-	}
-
-	/**
-	 * Checks for a Runtime_Check in a list of checks.
-	 *
-	 * @since n.e.x.t
-	 *
-	 * @param array $checks An array of Check instances.
-	 * @return bool True if a Runtime_Check exists in the array, false if not.
-	 */
-	private function has_runtime_check( array $checks ) {
-		foreach ( $checks as $check ) {
-			if ( $check instanceof Runtime_Check ) {
-				return true;
-			}
-		}
-
-		return false;
 	}
 }
